Rename `ActionBind` into `ActionBinding` and move into a dedicated module
Similar to `InputBinding`. Related methods also renamed.
