+++ src/input_condition.rs
@@ -28,7 +28,7 @@ pub const DEFAULT_ACTUATION: f32 = 0.5;
 /// or "release" events.
 ///
 /// Can be applied both to inputs and actions.
-/// See [`ActionBind::with_conditions`](super::actions::ActionBind::with_conditions)
+/// See [`ActionBinding::with_conditions`](super::action_binding::ActionBinding::with_conditions)
 /// and [`BindingBuilder::with_conditions`](super::input_binding::BindingBuilder::with_conditions).
 pub trait InputCondition: Sync + Send + Debug + 'static {
     /// Returns calculates state.
@@ -76,7 +76,7 @@ pub enum ConditionKind {
 }
 
 /// Conversion into iterator of bindings that could be passed into
-/// [`ActionBind::with_conditions`](super::actions::ActionBind::with_conditions)
+/// [`ActionBinding::with_conditions`](super::action_binding::ActionBinding::with_conditions)
 /// and [`BindingBuilder::with_conditions`](super::input_binding::BindingBuilder::with_conditions).
 pub trait IntoConditions {
     /// Returns an iterator over conditions.
