+++ src/input_modifier.rs
@@ -20,7 +20,7 @@ use crate::{action_value::ActionValue, actions::ActionsData};
 /// or changing how input maps to axes.
 ///
 /// Can be applied both to inputs and actions.
-/// See [`ActionBind::with_modifiers`](super::actions::ActionBind::with_modifiers)
+/// See [`ActionBinding::with_modifiers`](super::action_binding::ActionBinding::with_modifiers)
 /// and [`BindingBuilder::with_modifiers`](super::input_binding::BindingBuilder::with_modifiers).
 pub trait InputModifier: Sync + Send + Debug + 'static {
     /// Returns pre-processed value.
@@ -35,7 +35,7 @@ pub trait InputModifier: Sync + Send + Debug + 'static {
 }
 
 /// Conversion into iterator of bindings that could be passed into
-/// [`ActionBind::with_modifiers`](super::actions::ActionBind::with_modifiers)
+/// [`ActionBinding::with_modifiers`](super::action_binding::ActionBinding::with_modifiers)
 /// and [`BindingBuilder::with_modifiers`](super::input_binding::BindingBuilder::with_modifiers).
 pub trait IntoModifiers {
     /// Returns an iterator over modifiers.
